add: Filtered Enhancemen & Use Context for Sidebar

+++ src/app/devices/page.js
@@ -10,6 +10,7 @@ export default function Devices() {
     status: "Active",
   });
   const [editingIndex, setEditingIndex] = useState(null);
+  const [searchTerm, setSearchTerm] = useState("");
 
   useEffect(() => {
     const storedDevices = JSON.parse(localStorage.getItem("devices"));
@@ -25,9 +26,22 @@ export default function Devices() {
     }
   }, []);
 
+  const filteredDevices = devices.filter((device) =>
+    device.name.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
   const handleSaveDevice = (e) => {
     e.preventDefault();
     if (!newDevice.name || !newDevice.type) return;
+
+    const deviceExists = devices.some(
+      (device) => device.name === newDevice.name
+    );
+
+    if (deviceExists) {
+      alert("Bu isimde bir cihaz zaten var!");
+      return;
+    }
 
     let updatedDevices;
     if (editingIndex !== null) {
@@ -45,6 +59,11 @@ export default function Devices() {
   };
 
   const handleDeleteDevice = (index) => {
+    const confirmDelete = window.confirm(
+      "Bu cihazı silmek istediğinize emin misiniz?"
+    );
+    if (!confirmDelete) return;
+
     const updatedDevices = devices.filter((_, i) => i !== index);
     setDevices(updatedDevices);
     localStorage.setItem("devices", JSON.stringify(updatedDevices));
@@ -61,6 +80,13 @@ export default function Devices() {
 
       <div className="flex-1 flex flex-col items-center p-6 bg-gray-100 overflow-auto">
         <h1 className="text-3xl font-bold text-gray-700 mb-6">Cihazlarım</h1>
+        <input
+          type="text"
+          placeholder="Cihaz Ara..."
+          className="w-full max-w-md px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 mb-4"
+          value={searchTerm}
+          onChange={(e) => setSearchTerm(e.target.value)}
+        />
 
         <form
           onSubmit={handleSaveDevice}
@@ -131,7 +157,7 @@ export default function Devices() {
               </tr>
             </thead>
             <tbody>
-              {devices.map((device, index) => (
+              {filteredDevices.map((device, index) => (
                 <tr key={index} className="text-center">
                   <td className="border border-gray-300 p-3">{device.name}</td>
                   <td className="border border-gray-300 p-3">{device.type}</td>
